logging: split syslog functionality into separate file
This package is a transitive dependency in cilium-cli, but currently it
fails to build on platforms without syslog support (e.g. windows).
However, we want still to be able to build it for some these platforms,
see e.g. https://github.com/cilium/cilium-cli/issues/231. Fix the build
by moving all functionality specifiy to syslog into a separate file,
protected by build tags. The functionality will be stubbed out on
platforms without syslog.
